- Add the FNetCommand interface for the DEM_ data.
This is for helping any derived source port which have a different network system than ZDoom, like for example Zandronum, which may need to execute the demo command immediately.
Note: not every command is under this scheme, like DEM_(EMPTY)USERCMD, which is written directly to the stream in d_protocol, but it should be okay for the time being.