- Add the FNetCommand interface for the DEM_ data.
The new system works like this: when a new DEM_ command is created, Net_NewCommand must be used. After pushing all the parameters in the buffer (if needed), the command must be completed with Net_FinalizeCommand, so the whole buffer is pushed to the 'specials' ticcmds.
This is for helping any derived source port which have a different command buffer system than ZDoom, like for example Zandronum, which may need to execute the demo command immediately.
Note: not every command is under this scheme, like DEM_(EMPTY)USERCMD, which is written directly to the stream in d_protocol, but it should be okay for the time being.